FinalProject/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Insurance_Claim_Prediction"
version = "0.1.0"
description = "Production grade ML system for car insurance claims prediction"
authors = [{ name = "Sudhakar T" }]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "pydantic>=2.0.0",
    "onnxruntime>=1.16.0",
    "joblib>=1.2.0",
    "scikit-learn>=1.2.0",
    "mlflow>=3.1.4",
    "xgboost>=2.1.4",
    "skl2onnx>=1.20.0",
    "onnxmltools>=1.16.0",
    "pytest>=8.4.2",
    "httpx>=0.28.1",
    "black>=25.11.0",
    "isort>=6.1.0",
    "pre-commit>=4.3.0",
]

[project.optional-dependencies]
train = [
    "scikit-learn>=1.2.0",
    "xgboost>=2.0.0",
    "joblib>=1.2.0",
    "matplotlib>=3.9.4",
    "skl2onnx>=1.16.0",
    "onnxmltools>=1.12.0",
    "mlflow>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ml_system"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [
    "tests",
]
pythonpath = [
    ".",
]

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "yaml.*",
    "sklearn.*",
    "joblib.*",
    "onnxruntime.*",
    "skl2onnx.*",
    "onnxmltools.*",
    "mlflow.*",
]
ignore_missing_imports = true
